ImageUploading Web App
Create a REST API for saving image and label, image must be stored in
DB(base64 or
binary), listing of all images with labels, and on front­end create a
single page AngularJS app
which will consume APIs, ie functionality for uploading image and label,
and a list view for
diplsaying all images with associated labels. REST API must be done with
Node.js/Express or
with Laravel framework.